Current

+++ system/updates/to_run/engines/43/updater.sh
@@ -1,3 +1,5 @@
 #/bin/bash
-
+if test -f /opt/engines/run/utilities/fsconfigurator/running.yaml
+ then
  rm /opt/engines/run/utilities/fsconfigurator/running.yaml
+ fi
